refactor: move errors to memory pkg
