+++ pkg/storage/memory/errors.go
@@ -1,4 +1,4 @@
-package storage
+package memory
 
 import (
 	"errors"
@@ -6,8 +6,8 @@ import (
 	"reflect"
 )
 
-var ErrNotPointer = errors.New("fsm/storage: dest argument must be pointer")
-var ErrInvalidValue = errors.New("fsm/storage: dest value is nil or invalid")
+var ErrNotPointer = errors.New("fsm/storage/memory: dest argument must be pointer")
+var ErrInvalidValue = errors.New("fsm/storage/memory: dest value is nil or invalid")
 
 type ErrWrongTypeAssign struct {
 	Expect reflect.Type
