+++ pkg/storage/memory/memory_test.go
@@ -8,7 +8,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/vitaliy-ukiru/fsm-telebot/v2"
-	"github.com/vitaliy-ukiru/fsm-telebot/v2/pkg/storage"
 )
 
 func TestStorage_GetData(t *testing.T) {
@@ -52,7 +51,7 @@ func TestStorage_GetData(t *testing.T) {
 			data: m,
 			args: args{"right", false},
 			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorIs(t, err, storage.ErrNotPointer, i...)
+				return assert.ErrorIs(t, err, ErrNotPointer, i...)
 			},
 		},
 		{
@@ -60,7 +59,7 @@ func TestStorage_GetData(t *testing.T) {
 			data: m,
 			args: args{"foo", new(byte)},
 			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				var e *storage.ErrWrongTypeAssign
+				var e *ErrWrongTypeAssign
 				if !assert.ErrorAs(t, err, &e, i...) {
 					return false
 				}
@@ -74,7 +73,7 @@ func TestStorage_GetData(t *testing.T) {
 			data: m,
 			args: args{"age", (*int)(nil)},
 			wantErr: func(t assert.TestingT, err error, i ...interface{}) bool {
-				return assert.ErrorIs(t, err, storage.ErrInvalidValue, i...)
+				return assert.ErrorIs(t, err, ErrInvalidValue, i...)
 			},
 		},
 	}
